backend/push-notifications/requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
cryptography==41.0.7